aabb now works, enemies shows up, anemies get added in gamestate

+++ GameLibrary/Game/NCGame/Engine/animationcomponent.cpp
@@ -49,7 +49,7 @@ void AnimationComponent::Update()
 				m_sensei->SetState(Entity::DESTROY);
 				break;
 			case ePlayback::PING_PONG:
-				m_frame = Math::Clamp(m_frame, 0,(int)m_textures.size() - 1);
+				m_frame = Math::Clamp(m_frame, 0,m_textures.size() - 1);
 				m_direction = -m_direction;
 			}
